tr2/fmv: move to game/
