+++ src/tr2/meson.build
@@ -73,7 +73,6 @@ sources = [
   'config_map.c',
   'decomp/decomp.c',
   'decomp/flares.c',
-  'decomp/fmv.c',
   'decomp/savegame.c',
   'decomp/skidoo.c',
   'decomp/stats.c',
@@ -88,6 +87,7 @@ sources = [
   'game/demo.c',
   'game/effects.c',
   'game/fader.c',
+  'game/fmv.c',
   'game/game.c',
   'game/game_string.c',
   'game/gamebuf.c',
